+++ activerecord/lib/active_record/connection_adapters/mysql2_adapter.rb
@@ -572,12 +572,15 @@ module ActiveRecord
 
       # Returns a table's primary key and belonging sequence.
       def pk_and_sequence_for(table)
-        keys = []
-        result = execute("SHOW INDEX FROM #{quote_table_name(table)} WHERE Key_name = 'PRIMARY'", 'SCHEMA')
-        result.each(:symbolize_keys => true, :as => :hash) do |row|
-          keys << row[:Column_name]
+        result = execute("SHOW CREATE TABLE #{quote_table_name(table)}", 'SCHEMA')
+        create_table = result.first[1]
+
+        if create_table.to_s =~ /PRIMARY KEY\s+\((.+)\)/
+          keys = $1.split(",").map { |key| key.gsub(/`/, "") }
+          keys.length == 1 ? [keys.first, nil] : nil
+        else
+          nil
         end
-        keys.length == 1 ? [keys.first, nil] : nil
       end
 
       # Returns just a table's primary key