+++ activerecord/lib/active_record/connection_adapters/mysql_adapter.rb
@@ -755,13 +755,16 @@ module ActiveRecord
 
       # Returns a table's primary key and belonging sequence.
       def pk_and_sequence_for(table) #:nodoc:
-        keys = []
-        result = execute("SHOW INDEX FROM #{quote_table_name(table)} WHERE Key_name = 'PRIMARY'", 'SCHEMA')
-        result.each_hash do |h|
-          keys << h["Column_name"]
-        end
+        result = execute("SHOW CREATE TABLE #{quote_table_name(table)}", 'SCHEMA')
+        create_table = result.fetch_hash["Create Table"]
         result.free
-        keys.length == 1 ? [keys.first, nil] : nil
+
+        if create_table.to_s =~ /PRIMARY KEY\s+\((.+)\)/
+          keys = $1.split(",").map { |key| key.gsub(/`/, "") }
+          keys.length == 1 ? [keys.first, nil] : nil
+        else
+          nil
+        end
       end
 
       # Returns just a table's primary key